Massive refactoring.
The end goal is to make dope finally play nice as a WSGI app. It uses
the flask factory pattern (creation of app through a create_app()
method), and should be able to cope with multiple instances of the
app running on the same WSGI server.

At this time, only one instance is supported because the OpenID module
is not per-instance yet.